Move files back to the table_api folder
Also add a bit of documentation, and solve a signed / unsigned issue.
