+++ README.md
@@ -1,16 +1,37 @@
 Repo for storing code related to the foosball table at Telluride 2018.  AKA the best foosball table ever!
 
-## Flash instructions
-
-To flash the Arduino Nanos, open `table_api/slave/slave.ino` and edit the lines:
-```c
-const motion_type motion = translation; // one of 'translation', 'rotation'
-const position_type position = goalie; // one of 'goalie', 'defender', 'midfield', 'forward'
-```
-
-Master can be flashed without editing.
-
-
 ## Table API
 
-`test.py` shows how to use the `sensiball.py` library.
+### Requirements
+
+You must install *pyserial* to communicate with the Foosball table:
+```sh
+pip install pyserial
+```
+This package is known to interfere with the package *serial*. Uninstall the latter to avoid issues:
+```sh
+pip uninstall serial
+```
+
+### Importing the API from another directory
+
+To import the API from another directory, use the following code:
+```py
+import sys
+sys.path.insert(0, 'relative/path/to/table_api')
+from sensiball import Sensiball
+```
+
+### Examples
+
+*table_api/examples* provides several examples showing how to use the python API to communicate with the Foosball table.
+
+### Flashing the Arduinos
+
+- To flash the Arduino Nanos, open `table_api/slave/slave.ino` and edit the lines:
+```c
+#define SLAVE_ADDRESS 14 // 8/10/12/14 for Goalie/Defender/Midfield/Forward
+#define SLAVE_MODE    1  // 0/1 for Translation/Rotation
+```
+
+- To flash the Arduino Mega, open `table_api/master/master.ino`. Master can be flashed without editing.
